pyproject.toml
[project]
name = "token-context-analysis"
version = "0.1.0"
description = "Token and context window analysis for agentic traces stored in ClickHouse"
# readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clickhouse-connect>=0.7.0",
    "pydantic>=2.0",
    "google-adk>=1.0.0",
    "a2a-sdk==0.3.2",
    "litellm==1.82.1",
    "python-dotenv>=1.1.0",
    "loguru>=0.7.3",
]

[project.scripts]
token-analysis = "token_analysis.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/token_analysis", "src/agent", "src/server"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "e2e: end-to-end tests requiring a live ClickHouse instance",
]
